Homestay crud

+++ routes/web.php
@@ -2,9 +2,11 @@
 
 use App\Http\Controllers\Admin\AdminController;
 use App\Http\Controllers\Admin\MessageController;
+use App\Http\Controllers\Homestay\HomestayController;
 use App\Http\Controllers\PageController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Middleware\IsAdminMiddleware;
+use App\Http\Middleware\IsHomestayOwnerMiddleware;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', [PageController::class, 'index'])->name('page.home');
@@ -16,16 +18,20 @@ Route::get('/user/dashboard', function () {
     return view('user.dashboard');
 })->middleware(['auth', 'verified'])->name('user.dashboard');
 
+//Admin routes
 Route::get('/admin/dashboard', [AdminController::class, 'index'])->name('admin.dashboard')->middleware(['auth', IsAdminMiddleware::class]);
-
 Route::get('/admin/messages', [MessageController::class, 'index'])->name('messages.index')->middleware(['auth', IsAdminMiddleware::class]);
-Route::get('/admin/messages/{id}', [MessageController::class, 'show'])->name('messages.show');
-Route::delete('/admin/messages/{id}', [MessageController::class, 'destroy'])->name('messages.destroy');
+Route::get('/admin/messages/{id}', [MessageController::class, 'show'])->name('messages.show')->middleware(['auth', IsAdminMiddleware::class]);;
+Route::delete('/admin/messages/{id}', [MessageController::class, 'destroy'])->name('messages.destroy')->middleware(['auth', IsAdminMiddleware::class]);;
 
 
-Route::get('/homestay/dashboard', function () {
-    return view('homestay.dashboard');
-})->name('homestay.dashboard')->middleware('auth');
+// Homestay routes
+Route::get('/homestay/dashboard', [HomestayController::class, 'index'])->name('homestay.dashboard')->middleware('auth', IsHomestayOwnerMiddleware::class);
+Route::get('/homestay/inquiry', [HomestayController::class, 'inquiry'])->name('homestay.inquiry')->middleware('auth', IsHomestayOwnerMiddleware::class);
+Route::get('/homestay/create', [HomestayController::class, 'create'])->name('homestay.create')->middleware('auth', IsHomestayOwnerMiddleware::class);
+Route::post('/homestay/create', [HomestayController::class, 'store'])->name('homestay.store')->middleware('auth', IsHomestayOwnerMiddleware::class);
+Route::get('/homestay/{id}/edit', [HomestayController::class, 'edit'])->name('homestay.edit')->middleware('auth', IsHomestayOwnerMiddleware::class);
+Route::put('/homestay/{id}', [HomestayController::class, 'update'])->name('homestay.update')->middleware('auth', IsHomestayOwnerMiddleware::class);
 
 
 Route::middleware('auth')->group(function () {
